requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.22.0
plotly>=5.14.0
matplotlib>=3.7.1
pandas>=1.3.0
